+++ cmd/ecs-exporter/config_test.go
@@ -14,6 +14,8 @@ func TestConfigParse(t *testing.T) {
 		{true, []string{"--aws.region", "eu-west-1", "--web.listen-address", "0.0.0.0:9999"}},
 		{true, []string{"--aws.region", "eu-west-1"}},
 		{true, []string{"--aws.region", "eu-west-1", "--debug"}},
+		{true, []string{"--aws.region", "eu-west-1", "--aws.cluster-filter", ".*-prod-.*"}},
+		{false, []string{"--aws.region", "eu-west-1", "--aws.cluster-filter", "["}},
 		{false, []string{"--web.listen-address", "0.0.0.0:9999", "--web.telemetry-path", "/metrics2"}},
 		{false, []string{}},
 		{false, []string{"--aws-region", "eu-west-1"}},