+++ collector/collector_i_test.go
@@ -69,7 +69,7 @@ func TestCollectError(t *testing.T) {
 			},
 		}
 
-		exp, err := New("eu-west-1")
+		exp, err := New("eu-west-1", "")
 		if err != nil {
 			t.Errorf("Creation of exporter shouldn't error: %v", err)
 		}
@@ -108,96 +108,148 @@ func TestCollectError(t *testing.T) {
 func TestCollectOk(t *testing.T) {
 	tests := []struct {
 		cServices map[string][]*types.ECSService
+		cFilter   string
 		want      []string
+		dontWant  []string
 	}{
 		{
 			cServices: map[string][]*types.ECSService{
-				"c1uster1": {
+				"cluster1": {
 					&types.ECSService{ID: "s1", Name: "service1", DesiredT: 10, RunningT: 4, PendingT: 6}},
 			},
+			cFilter: ".*",
 			want: []string{
 				`ecs_up{region="eu-west-1"} 1`,
 				`ecs_cluster_total{region="eu-west-1"} 1`,
-				`ecs_service_total{cluster="c1uster1",region="eu-west-1"} 1`,
+				`ecs_service_total{cluster="cluster1",region="eu-west-1"} 1`,
 
-				`ecs_service_desired_tasks{cluster="c1uster1",region="eu-west-1",service="service1"} 10`,
-				`ecs_service_running_tasks{cluster="c1uster1",region="eu-west-1",service="service1"} 4`,
-				`ecs_service_pending_tasks{cluster="c1uster1",region="eu-west-1",service="service1"} 6`,
+				`ecs_service_desired_tasks{cluster="cluster1",region="eu-west-1",service="service1"} 10`,
+				`ecs_service_running_tasks{cluster="cluster1",region="eu-west-1",service="service1"} 4`,
+				`ecs_service_pending_tasks{cluster="cluster1",region="eu-west-1",service="service1"} 6`,
 			},
 		},
 		{
 			cServices: map[string][]*types.ECSService{
-				"c1uster1": {
+				"cluster1": {
 					&types.ECSService{ID: "s1", Name: "service1", DesiredT: 10, RunningT: 4, PendingT: 6},
 					&types.ECSService{ID: "s2", Name: "service2", DesiredT: 987, RunningT: 67, PendingT: 62},
 					&types.ECSService{ID: "s3", Name: "service3", DesiredT: 43, RunningT: 20, PendingT: 0},
 				},
 			},
+			cFilter: ".*",
 			want: []string{
 				`ecs_up{region="eu-west-1"} 1`,
 				`ecs_cluster_total{region="eu-west-1"} 1`,
-				`ecs_service_total{cluster="c1uster1",region="eu-west-1"} 3`,
+				`ecs_service_total{cluster="cluster1",region="eu-west-1"} 3`,
 
-				`ecs_service_desired_tasks{cluster="c1uster1",region="eu-west-1",service="service1"} 10`,
-				`ecs_service_running_tasks{cluster="c1uster1",region="eu-west-1",service="service1"} 4`,
-				`ecs_service_pending_tasks{cluster="c1uster1",region="eu-west-1",service="service1"} 6`,
+				`ecs_service_desired_tasks{cluster="cluster1",region="eu-west-1",service="service1"} 10`,
+				`ecs_service_running_tasks{cluster="cluster1",region="eu-west-1",service="service1"} 4`,
+				`ecs_service_pending_tasks{cluster="cluster1",region="eu-west-1",service="service1"} 6`,
 
-				`ecs_service_desired_tasks{cluster="c1uster1",region="eu-west-1",service="service2"} 987`,
-				`ecs_service_running_tasks{cluster="c1uster1",region="eu-west-1",service="service2"} 67`,
-				`ecs_service_pending_tasks{cluster="c1uster1",region="eu-west-1",service="service2"} 62`,
+				`ecs_service_desired_tasks{cluster="cluster1",region="eu-west-1",service="service2"} 987`,
+				`ecs_service_running_tasks{cluster="cluster1",region="eu-west-1",service="service2"} 67`,
+				`ecs_service_pending_tasks{cluster="cluster1",region="eu-west-1",service="service2"} 62`,
 
-				`ecs_service_desired_tasks{cluster="c1uster1",region="eu-west-1",service="service3"} 43`,
-				`ecs_service_running_tasks{cluster="c1uster1",region="eu-west-1",service="service3"} 20`,
-				`ecs_service_pending_tasks{cluster="c1uster1",region="eu-west-1",service="service3"} 0`,
+				`ecs_service_desired_tasks{cluster="cluster1",region="eu-west-1",service="service3"} 43`,
+				`ecs_service_running_tasks{cluster="cluster1",region="eu-west-1",service="service3"} 20`,
+				`ecs_service_pending_tasks{cluster="cluster1",region="eu-west-1",service="service3"} 0`,
 			},
 		},
 		{
 			cServices: map[string][]*types.ECSService{
-				"c1uster0": {&types.ECSService{ID: "s0", Name: "service0", DesiredT: 3, RunningT: 2, PendingT: 1}},
-				"c1uster1": {&types.ECSService{ID: "s0", Name: "service0", DesiredT: 10, RunningT: 5, PendingT: 5}},
-				"c1uster2": {&types.ECSService{ID: "s0", Name: "service0", DesiredT: 15, RunningT: 7, PendingT: 8}},
-				"c1uster3": {&types.ECSService{ID: "s0", Name: "service0", DesiredT: 30, RunningT: 15, PendingT: 15}},
-				"c1uster4": {&types.ECSService{ID: "s0", Name: "service0", DesiredT: 100, RunningT: 10, PendingT: 90}},
-				"c1uster5": {&types.ECSService{ID: "s0", Name: "service0", DesiredT: 75, RunningT: 50, PendingT: 25}},
+				"cluster0": {&types.ECSService{ID: "s0", Name: "service0", DesiredT: 3, RunningT: 2, PendingT: 1}},
+				"cluster1": {&types.ECSService{ID: "s0", Name: "service0", DesiredT: 10, RunningT: 5, PendingT: 5}},
+				"cluster2": {&types.ECSService{ID: "s0", Name: "service0", DesiredT: 15, RunningT: 7, PendingT: 8}},
+				"cluster3": {&types.ECSService{ID: "s0", Name: "service0", DesiredT: 30, RunningT: 15, PendingT: 15}},
+				"cluster4": {&types.ECSService{ID: "s0", Name: "service0", DesiredT: 100, RunningT: 10, PendingT: 90}},
+				"cluster5": {&types.ECSService{ID: "s0", Name: "service0", DesiredT: 75, RunningT: 50, PendingT: 25}},
 			},
+			cFilter: ".*",
 			want: []string{
 				`ecs_up{region="eu-west-1"} 1`,
 				`ecs_cluster_total{region="eu-west-1"} 6`,
-				`ecs_service_total{cluster="c1uster0",region="eu-west-1"} 1`,
-				`ecs_service_total{cluster="c1uster1",region="eu-west-1"} 1`,
-				`ecs_service_total{cluster="c1uster2",region="eu-west-1"} 1`,
-				`ecs_service_total{cluster="c1uster3",region="eu-west-1"} 1`,
-				`ecs_service_total{cluster="c1uster4",region="eu-west-1"} 1`,
-				`ecs_service_total{cluster="c1uster5",region="eu-west-1"} 1`,
+				`ecs_service_total{cluster="cluster0",region="eu-west-1"} 1`,
+				`ecs_service_total{cluster="cluster1",region="eu-west-1"} 1`,
+				`ecs_service_total{cluster="cluster2",region="eu-west-1"} 1`,
+				`ecs_service_total{cluster="cluster3",region="eu-west-1"} 1`,
+				`ecs_service_total{cluster="cluster4",region="eu-west-1"} 1`,
+				`ecs_service_total{cluster="cluster5",region="eu-west-1"} 1`,
 
-				`ecs_service_desired_tasks{cluster="c1uster0",region="eu-west-1",service="service0"} 3`,
-				`ecs_service_running_tasks{cluster="c1uster0",region="eu-west-1",service="service0"} 2`,
-				`ecs_service_pending_tasks{cluster="c1uster0",region="eu-west-1",service="service0"} 1`,
+				`ecs_service_desired_tasks{cluster="cluster0",region="eu-west-1",service="service0"} 3`,
+				`ecs_service_running_tasks{cluster="cluster0",region="eu-west-1",service="service0"} 2`,
+				`ecs_service_pending_tasks{cluster="cluster0",region="eu-west-1",service="service0"} 1`,
 
-				`ecs_service_desired_tasks{cluster="c1uster1",region="eu-west-1",service="service0"} 10`,
-				`ecs_service_running_tasks{cluster="c1uster1",region="eu-west-1",service="service0"} 5`,
-				`ecs_service_pending_tasks{cluster="c1uster1",region="eu-west-1",service="service0"} 5`,
+				`ecs_service_desired_tasks{cluster="cluster1",region="eu-west-1",service="service0"} 10`,
+				`ecs_service_running_tasks{cluster="cluster1",region="eu-west-1",service="service0"} 5`,
+				`ecs_service_pending_tasks{cluster="cluster1",region="eu-west-1",service="service0"} 5`,
 
-				`ecs_service_desired_tasks{cluster="c1uster2",region="eu-west-1",service="service0"} 15`,
-				`ecs_service_running_tasks{cluster="c1uster2",region="eu-west-1",service="service0"} 7`,
-				`ecs_service_pending_tasks{cluster="c1uster2",region="eu-west-1",service="service0"} 8`,
+				`ecs_service_desired_tasks{cluster="cluster2",region="eu-west-1",service="service0"} 15`,
+				`ecs_service_running_tasks{cluster="cluster2",region="eu-west-1",service="service0"} 7`,
+				`ecs_service_pending_tasks{cluster="cluster2",region="eu-west-1",service="service0"} 8`,
 
-				`ecs_service_desired_tasks{cluster="c1uster3",region="eu-west-1",service="service0"} 30`,
-				`ecs_service_running_tasks{cluster="c1uster3",region="eu-west-1",service="service0"} 15`,
-				`ecs_service_pending_tasks{cluster="c1uster3",region="eu-west-1",service="service0"} 15`,
+				`ecs_service_desired_tasks{cluster="cluster3",region="eu-west-1",service="service0"} 30`,
+				`ecs_service_running_tasks{cluster="cluster3",region="eu-west-1",service="service0"} 15`,
+				`ecs_service_pending_tasks{cluster="cluster3",region="eu-west-1",service="service0"} 15`,
 
-				`ecs_service_desired_tasks{cluster="c1uster4",region="eu-west-1",service="service0"} 100`,
-				`ecs_service_running_tasks{cluster="c1uster4",region="eu-west-1",service="service0"} 10`,
-				`ecs_service_pending_tasks{cluster="c1uster4",region="eu-west-1",service="service0"} 90`,
+				`ecs_service_desired_tasks{cluster="cluster4",region="eu-west-1",service="service0"} 100`,
+				`ecs_service_running_tasks{cluster="cluster4",region="eu-west-1",service="service0"} 10`,
+				`ecs_service_pending_tasks{cluster="cluster4",region="eu-west-1",service="service0"} 90`,
 
-				`ecs_service_desired_tasks{cluster="c1uster5",region="eu-west-1",service="service0"} 75`,
-				`ecs_service_running_tasks{cluster="c1uster5",region="eu-west-1",service="service0"} 50`,
-				`ecs_service_pending_tasks{cluster="c1uster5",region="eu-west-1",service="service0"} 25`,
+				`ecs_service_desired_tasks{cluster="cluster5",region="eu-west-1",service="service0"} 75`,
+				`ecs_service_running_tasks{cluster="cluster5",region="eu-west-1",service="service0"} 50`,
+				`ecs_service_pending_tasks{cluster="cluster5",region="eu-west-1",service="service0"} 25`,
 			},
 		},
 		{
 			cServices: map[string][]*types.ECSService{
-				"c1uster1": {
+				"cluster0": {&types.ECSService{ID: "s0", Name: "service0", DesiredT: 3, RunningT: 2, PendingT: 1}},
+				"cluster1": {&types.ECSService{ID: "s0", Name: "service0", DesiredT: 10, RunningT: 5, PendingT: 5}},
+				"cluster2": {&types.ECSService{ID: "s0", Name: "service0", DesiredT: 15, RunningT: 7, PendingT: 8}},
+				"cluster3": {&types.ECSService{ID: "s0", Name: "service0", DesiredT: 30, RunningT: 15, PendingT: 15}},
+				"cluster4": {&types.ECSService{ID: "s0", Name: "service0", DesiredT: 100, RunningT: 10, PendingT: 90}},
+				"cluster5": {&types.ECSService{ID: "s0", Name: "service0", DesiredT: 75, RunningT: 50, PendingT: 25}},
+			},
+			cFilter: "cluster[024]",
+			want: []string{
+				`ecs_up{region="eu-west-1"} 1`,
+				`ecs_cluster_total{region="eu-west-1"} 6`,
+				`ecs_service_total{cluster="cluster0",region="eu-west-1"} 1`,
+				`ecs_service_total{cluster="cluster2",region="eu-west-1"} 1`,
+				`ecs_service_total{cluster="cluster4",region="eu-west-1"} 1`,
+
+				`ecs_service_desired_tasks{cluster="cluster0",region="eu-west-1",service="service0"} 3`,
+				`ecs_service_running_tasks{cluster="cluster0",region="eu-west-1",service="service0"} 2`,
+				`ecs_service_pending_tasks{cluster="cluster0",region="eu-west-1",service="service0"} 1`,
+
+				`ecs_service_desired_tasks{cluster="cluster2",region="eu-west-1",service="service0"} 15`,
+				`ecs_service_running_tasks{cluster="cluster2",region="eu-west-1",service="service0"} 7`,
+				`ecs_service_pending_tasks{cluster="cluster2",region="eu-west-1",service="service0"} 8`,
+
+				`ecs_service_desired_tasks{cluster="cluster4",region="eu-west-1",service="service0"} 100`,
+				`ecs_service_running_tasks{cluster="cluster4",region="eu-west-1",service="service0"} 10`,
+				`ecs_service_pending_tasks{cluster="cluster4",region="eu-west-1",service="service0"} 90`,
+			},
+			dontWant: []string{
+				`ecs_service_total{cluster="cluster1",region="eu-west-1"} 1`,
+				`ecs_service_total{cluster="cluster3",region="eu-west-1"} 1`,
+				`ecs_service_total{cluster="cluster5",region="eu-west-1"} 1`,
+
+				`ecs_service_desired_tasks{cluster="cluster1",region="eu-west-1",service="service0"} 10`,
+				`ecs_service_running_tasks{cluster="cluster1",region="eu-west-1",service="service0"} 5`,
+				`ecs_service_pending_tasks{cluster="cluster1",region="eu-west-1",service="service0"} 5`,
+
+				`ecs_service_desired_tasks{cluster="cluster3",region="eu-west-1",service="service0"} 30`,
+				`ecs_service_running_tasks{cluster="cluster3",region="eu-west-1",service="service0"} 15`,
+				`ecs_service_pending_tasks{cluster="cluster3",region="eu-west-1",service="service0"} 15`,
+
+				`ecs_service_desired_tasks{cluster="cluster5",region="eu-west-1",service="service0"} 75`,
+				`ecs_service_running_tasks{cluster="cluster5",region="eu-west-1",service="service0"} 50`,
+				`ecs_service_pending_tasks{cluster="cluster5",region="eu-west-1",service="service0"} 25`,
+			},
+		},
+		{
+			cServices: map[string][]*types.ECSService{
+				"cluster1": {
 					&types.ECSService{ID: "s1", Name: "service1", DesiredT: 10, RunningT: 4, PendingT: 6},
 					&types.ECSService{ID: "s2", Name: "service2", DesiredT: 987, RunningT: 67, PendingT: 62},
 					&types.ECSService{ID: "s3", Name: "service3", DesiredT: 43, RunningT: 20, PendingT: 0},
@@ -205,58 +257,61 @@ func TestCollectOk(t *testing.T) {
 					&types.ECSService{ID: "s5", Name: "service5", DesiredT: 3, RunningT: 2, PendingT: 1},
 				},
 
-				"c1uster2": {
+				"cluster2": {
 					&types.ECSService{ID: "s98", Name: "service98", DesiredT: 100, RunningT: 50, PendingT: 23},
 				},
 
-				"c1uster3": {
+				"cluster3": {
 					&types.ECSService{ID: "s1000", Name: "service1000", DesiredT: 1000, RunningT: 500, PendingT: 500},
 					&types.ECSService{ID: "s2000", Name: "service2000", DesiredT: 2000, RunningT: 1997, PendingT: 3},
 					&types.ECSService{ID: "s3000", Name: "service3000", DesiredT: 3000, RunningT: 2000, PendingT: 1000},
 				},
 			},
+			cFilter: "^cluster[^2]$",
 			want: []string{
 				`ecs_up{region="eu-west-1"} 1`,
 				`ecs_cluster_total{region="eu-west-1"} 3`,
-				`ecs_service_total{cluster="c1uster1",region="eu-west-1"} 5`,
-				`ecs_service_total{cluster="c1uster2",region="eu-west-1"} 1`,
-				`ecs_service_total{cluster="c1uster3",region="eu-west-1"} 3`,
+				`ecs_service_total{cluster="cluster1",region="eu-west-1"} 5`,
+				`ecs_service_total{cluster="cluster3",region="eu-west-1"} 3`,
 
-				`ecs_service_desired_tasks{cluster="c1uster1",region="eu-west-1",service="service1"} 10`,
-				`ecs_service_running_tasks{cluster="c1uster1",region="eu-west-1",service="service1"} 4`,
-				`ecs_service_pending_tasks{cluster="c1uster1",region="eu-west-1",service="service1"} 6`,
+				`ecs_service_desired_tasks{cluster="cluster1",region="eu-west-1",service="service1"} 10`,
+				`ecs_service_running_tasks{cluster="cluster1",region="eu-west-1",service="service1"} 4`,
+				`ecs_service_pending_tasks{cluster="cluster1",region="eu-west-1",service="service1"} 6`,
 
-				`ecs_service_desired_tasks{cluster="c1uster1",region="eu-west-1",service="service2"} 987`,
-				`ecs_service_running_tasks{cluster="c1uster1",region="eu-west-1",service="service2"} 67`,
-				`ecs_service_pending_tasks{cluster="c1uster1",region="eu-west-1",service="service2"} 62`,
+				`ecs_service_desired_tasks{cluster="cluster1",region="eu-west-1",service="service2"} 987`,
+				`ecs_service_running_tasks{cluster="cluster1",region="eu-west-1",service="service2"} 67`,
+				`ecs_service_pending_tasks{cluster="cluster1",region="eu-west-1",service="service2"} 62`,
 
-				`ecs_service_desired_tasks{cluster="c1uster1",region="eu-west-1",service="service3"} 43`,
-				`ecs_service_running_tasks{cluster="c1uster1",region="eu-west-1",service="service3"} 20`,
-				`ecs_service_pending_tasks{cluster="c1uster1",region="eu-west-1",service="service3"} 0`,
+				`ecs_service_desired_tasks{cluster="cluster1",region="eu-west-1",service="service3"} 43`,
+				`ecs_service_running_tasks{cluster="cluster1",region="eu-west-1",service="service3"} 20`,
+				`ecs_service_pending_tasks{cluster="cluster1",region="eu-west-1",service="service3"} 0`,
 
-				`ecs_service_desired_tasks{cluster="c1uster1",region="eu-west-1",service="service4"} 88`,
-				`ecs_service_running_tasks{cluster="c1uster1",region="eu-west-1",service="service4"} 77`,
-				`ecs_service_pending_tasks{cluster="c1uster1",region="eu-west-1",service="service4"} 11`,
+				`ecs_service_desired_tasks{cluster="cluster1",region="eu-west-1",service="service4"} 88`,
+				`ecs_service_running_tasks{cluster="cluster1",region="eu-west-1",service="service4"} 77`,
+				`ecs_service_pending_tasks{cluster="cluster1",region="eu-west-1",service="service4"} 11`,
 
-				`ecs_service_desired_tasks{cluster="c1uster1",region="eu-west-1",service="service5"} 3`,
-				`ecs_service_running_tasks{cluster="c1uster1",region="eu-west-1",service="service5"} 2`,
-				`ecs_service_pending_tasks{cluster="c1uster1",region="eu-west-1",service="service5"} 1`,
+				`ecs_service_desired_tasks{cluster="cluster1",region="eu-west-1",service="service5"} 3`,
+				`ecs_service_running_tasks{cluster="cluster1",region="eu-west-1",service="service5"} 2`,
+				`ecs_service_pending_tasks{cluster="cluster1",region="eu-west-1",service="service5"} 1`,
 
-				`ecs_service_desired_tasks{cluster="c1uster2",region="eu-west-1",service="service98"} 100`,
-				`ecs_service_running_tasks{cluster="c1uster2",region="eu-west-1",service="service98"} 50`,
-				`ecs_service_pending_tasks{cluster="c1uster2",region="eu-west-1",service="service98"} 23`,
+				`ecs_service_desired_tasks{cluster="cluster3",region="eu-west-1",service="service1000"} 1000`,
+				`ecs_service_running_tasks{cluster="cluster3",region="eu-west-1",service="service1000"} 500`,
+				`ecs_service_pending_tasks{cluster="cluster3",region="eu-west-1",service="service1000"} 500`,
 
-				`ecs_service_desired_tasks{cluster="c1uster3",region="eu-west-1",service="service1000"} 1000`,
-				`ecs_service_running_tasks{cluster="c1uster3",region="eu-west-1",service="service1000"} 500`,
-				`ecs_service_pending_tasks{cluster="c1uster3",region="eu-west-1",service="service1000"} 500`,
+				`ecs_service_desired_tasks{cluster="cluster3",region="eu-west-1",service="service2000"} 2000`,
+				`ecs_service_running_tasks{cluster="cluster3",region="eu-west-1",service="service2000"} 1997`,
+				`ecs_service_pending_tasks{cluster="cluster3",region="eu-west-1",service="service2000"} 3`,
 
-				`ecs_service_desired_tasks{cluster="c1uster3",region="eu-west-1",service="service2000"} 2000`,
-				`ecs_service_running_tasks{cluster="c1uster3",region="eu-west-1",service="service2000"} 1997`,
-				`ecs_service_pending_tasks{cluster="c1uster3",region="eu-west-1",service="service2000"} 3`,
+				`ecs_service_desired_tasks{cluster="cluster3",region="eu-west-1",service="service3000"} 3000`,
+				`ecs_service_running_tasks{cluster="cluster3",region="eu-west-1",service="service3000"} 2000`,
+				`ecs_service_pending_tasks{cluster="cluster3",region="eu-west-1",service="service3000"} 1000`,
+			},
+			dontWant: []string{
+				`ecs_service_total{cluster="cluster2",region="eu-west-1"} 1`,
 
-				`ecs_service_desired_tasks{cluster="c1uster3",region="eu-west-1",service="service3000"} 3000`,
-				`ecs_service_running_tasks{cluster="c1uster3",region="eu-west-1",service="service3000"} 2000`,
-				`ecs_service_pending_tasks{cluster="c1uster3",region="eu-west-1",service="service3000"} 1000`,
+				`ecs_service_desired_tasks{cluster="cluster2",region="eu-west-1",service="service98"} 100`,
+				`ecs_service_running_tasks{cluster="cluster2",region="eu-west-1",service="service98"} 50`,
+				`ecs_service_pending_tasks{cluster="cluster2",region="eu-west-1",service="service98"} 23`,
 			},
 		},
 	}
@@ -265,7 +320,7 @@ func TestCollectOk(t *testing.T) {
 
 		e := &ECSMockClient{sd: test.cServices}
 
-		exp, err := New("eu-west-1")
+		exp, err := New("eu-west-1", test.cFilter)
 		if err != nil {
 			t.Errorf("Creation of exporter shouldn't error: %v", err)
 		}
@@ -287,6 +342,12 @@ func TestCollectOk(t *testing.T) {
 		for _, m := range test.want {
 			if !strings.Contains(got, m) {
 				t.Errorf("%+v\n -Expected metric data but missing: %s", test, m)
+			}
+		}
+
+		for _, m := range test.dontWant {
+			if strings.Contains(got, m) {
+				t.Errorf("%+v\n -Didn't expected metric data but found: %s", test, m)
 			}
 		}